base: Add interface for installers that are aware of many bundles
These installers are always run, even if the bundle they are from is
not included in the install command.

For example, a bundle-aware installer could read a file from the
Resources/ folder of every bundle and do something with it, e.g. seed
the database.

However, installers implementing this interface must only effect the
supplied bundles.

For example, if a FooInstaller from the FooBundle was bundle-aware,
and the install command was ran specifying the BarBundle and
BazBundle, the FooInstaller must only execute for the Bar and Baz
bundle, NOT the FooBundle.