base: Separate action isGranted from display logic
isGranted is solely for deciding if an action should run on an entity.

isButtonAvailable is used for deciding to display the action button
for each entity.

isBatchOptionAvailable is used for deciding if the action should be
displayed in the batchActions dropdown.

All of these methods are now options passed to ActionConfig#add() and
addInstance(), not methods on ActionInterface. This makes it very easy
to override permission and display logic without creating a new action
class.